+++ Entities/ACraft.cpp
@@ -265,7 +265,7 @@ void ACraft::Clear()
     m_HatchDelay = 0;
     m_HatchOpenSound = nullptr;
     m_HatchCloseSound = nullptr;
-    m_NewInventory.clear();
+    m_CollectedInventory.clear();
     m_Exits.clear();
     m_CurrentExit = m_Exits.begin();
     m_ExitInterval = 1000;
@@ -321,8 +321,8 @@ int ACraft::Create(const ACraft &reference)
 	} else if (reference.m_HatchOpenSound) {
 		m_HatchCloseSound = dynamic_cast<SoundContainer *>(reference.m_HatchOpenSound->Clone());
 	}
-	for (deque<MovableObject *>::const_iterator niItr = reference.m_NewInventory.begin(); niItr != reference.m_NewInventory.end(); ++niItr)
-        m_NewInventory.push_back(dynamic_cast<MovableObject *>((*niItr)->Clone()));
+	for (deque<MovableObject *>::const_iterator niItr = reference.m_CollectedInventory.begin(); niItr != reference.m_CollectedInventory.end(); ++niItr)
+        m_CollectedInventory.push_back(dynamic_cast<MovableObject *>((*niItr)->Clone()));
     for (list<Exit>::const_iterator eItr = reference.m_Exits.begin(); eItr != reference.m_Exits.end(); ++eItr)
         m_Exits.push_back(*eItr);
     m_CurrentExit = m_Exits.begin();
@@ -450,7 +450,7 @@ float ACraft::GetTotalValue(int nativeModule, float foreignMult, float nativeMul
     float totalValue = Actor::GetTotalValue(nativeModule, foreignMult, nativeMult);
 
     MOSprite *pItem = 0;
-    for (deque<MovableObject *>::const_iterator itr = m_NewInventory.begin(); itr != m_NewInventory.end(); ++itr)
+    for (deque<MovableObject *>::const_iterator itr = m_CollectedInventory.begin(); itr != m_CollectedInventory.end(); ++itr)
     {
         pItem = dynamic_cast<MOSprite *>(*itr);
         if (pItem)
@@ -472,7 +472,7 @@ bool ACraft::HasObject(string objectName) const
     if (Actor::HasObject(objectName))
         return true;
 
-    for (deque<MovableObject *>::const_iterator itr = m_NewInventory.begin(); itr != m_NewInventory.end(); ++itr)
+    for (deque<MovableObject *>::const_iterator itr = m_CollectedInventory.begin(); itr != m_CollectedInventory.end(); ++itr)
     {
         if ((*itr) && (*itr)->HasObject(objectName))
             return true;
@@ -494,7 +494,7 @@ bool ACraft::HasObjectInGroup(std::string groupName) const
     if (Actor::HasObjectInGroup(groupName))
         return true;
 
-    for (deque<MovableObject *>::const_iterator itr = m_NewInventory.begin(); itr != m_NewInventory.end(); ++itr)
+    for (deque<MovableObject *>::const_iterator itr = m_CollectedInventory.begin(); itr != m_CollectedInventory.end(); ++itr)
     {
         if ((*itr) && (*itr)->HasObjectInGroup(groupName))
             return true;
@@ -515,7 +515,7 @@ void ACraft::SetTeam(int team)
 
     // Also set all actors in the new inventory
     Actor *pActor = 0;
-    for (deque<MovableObject *>::iterator itr = m_NewInventory.begin(); itr != m_NewInventory.end(); ++itr)
+    for (deque<MovableObject *>::iterator itr = m_CollectedInventory.begin(); itr != m_CollectedInventory.end(); ++itr)
     {
         pActor = dynamic_cast<Actor *>(*itr);
         if (pActor)
@@ -619,11 +619,11 @@ void ACraft::CloseHatch()
         m_HatchTimer.Reset();
 
         // When closing, move all newly added inventory to the regular inventory list so it'll be ejected next time doors open
-        for (deque<MovableObject *>::const_iterator niItr = m_NewInventory.begin(); niItr != m_NewInventory.end(); ++niItr)
+        for (deque<MovableObject *>::const_iterator niItr = m_CollectedInventory.begin(); niItr != m_CollectedInventory.end(); ++niItr)
             m_Inventory.push_back(*niItr);
 
         // Clear the new inventory hold, it's all been moved to the regular inventory
-        m_NewInventory.clear();
+        m_CollectedInventory.clear();
 
         // PSCHHT
 		if (m_HatchCloseSound) { m_HatchCloseSound->Play(m_Pos); }
@@ -645,7 +645,7 @@ void ACraft::AddInventoryItem(MovableObject *pItemToAdd)
         // If the hatch is open, then only add the new item to the intermediate new inventory list
         // so that it doesn't get chucked out right away again
         if (m_HatchState == OPEN || m_HatchState == OPENING)
-            m_NewInventory.push_back(pItemToAdd);
+            m_CollectedInventory.push_back(pItemToAdd);
         // If doors are already closed, it's safe to put the item directly the regular inventory
         else
             m_Inventory.push_back(pItemToAdd);
@@ -792,6 +792,17 @@ void ACraft::DropAllInventory()
     }
 }
 
+/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+float ACraft::GetCollectedInventoryMass() const {
+	float inventoryMass = 0.0F;
+	for (const MovableObject *inventoryItem : m_CollectedInventory) {
+		inventoryMass += inventoryItem->GetMass();
+	}
+	return inventoryMass;
+}
+
+/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 //////////////////////////////////////////////////////////////////////////////////////////
 // Virtual method:  OnMOHit
@@ -983,22 +994,21 @@ void ACraft::Update()
 // Description:     Draws this Actor's current graphical HUD overlay representation to a
 //                  BITMAP of choice.
 
-void ACraft::DrawHUD(BITMAP *pTargetBitmap, const Vector &targetPos, int whichScreen, bool playerControlled)
-{
+void ACraft::DrawHUD(BITMAP *pTargetBitmap, const Vector &targetPos, int whichScreen, bool playerControlled) {
+	m_HUDStack = -m_CharHeight / 2;
+
     if (!m_HUDVisible)
         return;
 
     // Only do HUD if on a team
     if (m_Team < 0)
         return;
-
-    // Only draw if the team viewing this is on the same team OR has seen the space where this is located
-    int viewingTeam = g_ActivityMan.GetActivity()->GetTeamOfPlayer(g_ActivityMan.GetActivity()->PlayerOfScreen(whichScreen));
-    if (viewingTeam != m_Team && viewingTeam != Activity::NoTeam)
-    {
-        if (g_SceneMan.IsUnseen(m_Pos.m_X, m_Pos.m_Y, viewingTeam))
-            return;
-    }
+	
+	// Only draw if the team viewing this is on the same team OR has seen the space where this is located.
+	int viewingTeam = g_ActivityMan.GetActivity()->GetTeamOfPlayer(g_ActivityMan.GetActivity()->PlayerOfScreen(whichScreen));
+	if (viewingTeam != m_Team && viewingTeam != Activity::NoTeam && (!g_SettingsMan.ShowEnemyHUD() || g_SceneMan.IsUnseen(m_Pos.GetFloorIntX(), m_Pos.GetFloorIntY(), viewingTeam))) {
+		return;
+	}
 
     Actor::DrawHUD(pTargetBitmap, targetPos, whichScreen);